+++ src/frontend/components/SessionSelectionModal/index.tsx
@@ -19,115 +19,7 @@ import {
   useSetOpenSession,
   useUpsertSession,
 } from 'src/frontend/hooks/useSession';
-
-type SessionOption = {
-  label: string;
-  value: string;
-  selected?: boolean;
-  disabled?: boolean;
-};
-
-type SessionSelectionFormProps = {
-  isFirstTime: boolean;
-  options: SessionOption[];
-};
-
-export function SessionSelectionForm(props: SessionSelectionFormProps) {
-  const { options } = props;
-  const navigate = useNavigate();
-  const { mutateAsync: setOpenSession } = useSetOpenSession();
-  const { mutateAsync: upsertSession } = useUpsertSession();
-
-  const onCreateNewSession = async (formEl: HTMLElement) => {
-    const newSessionName = (formEl.querySelector('input') as HTMLInputElement).value;
-
-    const newSession = await upsertSession({
-      id: getRandomSessionId(),
-      name: newSessionName,
-    });
-
-    const newSessionId = newSession.id;
-
-    // set the new session id;
-    await setOpenSession(newSessionId);
-
-    // go back to homepage before switching session
-    navigate('/', { replace: true });
-
-    // then set it as current session
-    await setCurrentSessionId(newSessionId);
-  };
-
-  const onSelectSession = async (newSessionId: string) => {
-    // set the new session id;
-    await setOpenSession(newSessionId);
-
-    // go back to homepage before switching session
-    navigate('/', { replace: true });
-
-    // then set it as current session
-    await setCurrentSessionId(newSessionId);
-  };
-
-  let defaultSessionName =
-    options.length === 0 ? `New Session ${new Date().toLocaleDateString()}` : '';
-
-  return (
-    <Box sx={{ display: 'flex', flexDirection: 'column', gap: 1 }}>
-      <Box>Please select a session from below:</Box>
-
-      <List>
-        {options.map((option) => {
-          const onSelectThisSession = () => {
-            if (!option.disabled) {
-              // don't let them select already selected session
-              onSelectSession(option.value);
-            }
-          };
-          const labelId = `session-option-${option.value}`;
-          return (
-            <ListItem
-              button
-              key={option.value}
-              disabled={option.disabled}
-              selected={option.selected}
-              onClick={onSelectThisSession}>
-              <ListItemIcon>
-                <Checkbox
-                  edge='start'
-                  checked={!!option.selected}
-                  tabIndex={-1}
-                  inputProps={{ 'aria-labelledby': labelId }}
-                />
-              </ListItemIcon>
-              <ListItemText id={labelId} primary={option.label} />
-            </ListItem>
-          );
-        })}
-      </List>
-
-      <form
-        onSubmit={(e) => {
-          e.preventDefault();
-          onCreateNewSession(e.target as HTMLElement);
-        }}>
-        <Box sx={{ display: 'flex', gap: 1, alignItems: 'center' }}>
-          <TextField
-            placeholder='Enter name for the new session'
-            label='New Session Name'
-            size='small'
-            required
-            sx={{ flexGrow: 1 }}
-            defaultValue={defaultSessionName}
-          />
-          <Button type='submit' size='small'>
-            Create Session
-          </Button>
-        </Box>
-      </form>
-    </Box>
-  );
-}
+import SessionSelectionForm from 'src/frontend/components/SessionSelectionForm';
 
 export default function SessionSelectionModal() {
   const navigate = useNavigate();
